pyproject.toml
[project]
name = "python-executor-mcp"
version = "0.1.0"
description = "A cross-platform Python code executor with MCP server support for LLM integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.0.0",
    "python-dotenv>=1.1.0",
]

[project.optional-dependencies]
full = [
    # AWS
    "boto3>=1.42.4",
    # Security
    "cryptography>=46.0.3",
    # Date/Time
    "dateparser>=1.2.2",
    "arrow>=1.4.0",
    "pytz>=2025.2",
    # Microsoft
    "exchangelib>=5.6.0",
    "msal>=1.34.0",
    # Data Processing
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "orjson>=3.11.5",
    "pydantic>=2.12.5",
    # Excel/Office
    "openpyxl>=3.1.5",
    "xlrd>=2.0.2",
    "xlsxwriter>=3.2.9",
    "python-docx>=1.2.0",
    "pypdf>=6.4.1",
    # Database
    "pymysql>=1.1.2",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.44",
    # HTTP/Network
    "requests>=2.32.5",
    "httpx>=0.28.1",
    "aiohttp>=3.13.2",
    "paramiko>=4.0.0",
    # Web Scraping
    "beautifulsoup4>=4.14.3",
    "lxml>=6.0.2",
    "chardet>=5.2.0",
    # Text/Markdown
    "markdown>=3.10",
    "jinja2>=3.1.6",
    "pyyaml>=6.0.3",
    # Image
    "pillow>=12.0.0",
    "matplotlib>=3.10.7",
    # CLI/Display
    "rich>=14.2.0",
    "tabulate>=0.9.0",
    "tqdm>=4.67.1",
    # Utilities
    "pyperclip>=1.11.0",
    "wrapt>=2.0.1",
    # Windows only
    "pywin32>=311; sys_platform == 'win32'",
]

[project.scripts]
python-executor = "pyx_cli.cli:main"
pyx = "pyx_cli.cli:main"
python-executor-mcp = "pyx_mcp.server:main"
pyx-mcp = "pyx_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pyx_core", "src/pyx_cli", "src/pyx_mcp"]
